requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
requests==2.32.3
confluent_kafka==2.6.1
fastavro==1.9.7
aws-msk-iam-sasl-signer-python==1.0.1
